+++ ormar/__init__.py
@@ -75,7 +75,7 @@ class UndefinedType:  # pragma no cover
 
 Undefined = UndefinedType()
 
-__version__ = "0.10.2"
+__version__ = "0.10.3"
 __all__ = [
     "Integer",
     "BigInteger",