+++ ormar/fields/foreign_key.py
@@ -233,8 +233,12 @@ def ForeignKey(  # noqa CFQ002
 
     owner = kwargs.pop("owner", None)
     self_reference = kwargs.pop("self_reference", False)
+
     orders_by = kwargs.pop("orders_by", None)
     related_orders_by = kwargs.pop("related_orders_by", None)
+
+    skip_reverse = kwargs.pop("skip_reverse", False)
+    skip_field = kwargs.pop("skip_field", False)
 
     validate_not_allowed_fields(kwargs)
 
@@ -274,6 +278,8 @@ def ForeignKey(  # noqa CFQ002
         is_relation=True,
         orders_by=orders_by,
         related_orders_by=related_orders_by,
+        skip_reverse=skip_reverse,
+        skip_field=skip_field,
     )
 
     Field = type("ForeignKey", (ForeignKeyField, BaseField), {})
@@ -311,6 +317,30 @@ class ForeignKeyField(BaseField):
         :rtype: str
         """
         return self.related_name or self.owner.get_name() + "s"
+
+    def default_target_field_name(self, reverse: bool = False) -> str:
+        """
+        Returns default target model name on through model.
+        :param reverse: flag to grab name without accessing related field
+        :type reverse: bool
+        :return: name of the field
+        :rtype: str
+        """
+        self_rel_prefix = "from_" if not reverse else "to_"
+        prefix = self_rel_prefix if self.self_reference else ""
+        return f"{prefix}{self.to.get_name()}"
+
+    def default_source_field_name(self, reverse: bool = False) -> str:
+        """
+        Returns default target model name on through model.
+        :param reverse: flag to grab name without accessing related field
+        :type reverse: bool
+        :return: name of the field
+        :rtype: str
+        """
+        self_rel_prefix = "to_" if not reverse else "from_"
+        prefix = self_rel_prefix if self.self_reference else ""
+        return f"{prefix}{self.owner.get_name()}"
 
     def evaluate_forward_ref(self, globalns: Any, localns: Any) -> None:
         """